Husks now teleport to owners location, Player list extended
This is a workaround to the stealth detection problem whereby passing units such as bikes would reveal the husks health bar while being captured.
To mitigate this I now teleport the husk to the start location of the owner. Furthermore, I have extended the team count to factor in campaign teams and missions.